Product Archiving
It is now possible to archive contracts.
This means that instead of saving references to products and clients, the contract now stores a copy of the product.

The advantage now is that products can now be edited without the fear of corrupting an already existing contract.

Before saving a product, it is now being checked, whether it could potentially interfere with any given contract. The user is then asked to either "SAVE ANYWAY", which risks corruption but may be something worth risking if you are pretty sure, or "ARCHIVE ALL", which will archive all contracts that belong to the product.