Globalization
Implemented multi-language support by using a LanguageProvider. It fetches the assets/Texts.json and can provide a text based on the current language.

For example: 'Buttons.Save' would be the path to the save button.
If the language was 'de', it would return "Speichern", if it was 'en', it would return "Save"