Conditional pricing
Conditianol prices are prices that are affected by choices. Each choice has a cond field that can take a number.
To use that number, you can specify its usage in a components price.

Usage:
Price set to:
cond:[ADDRESS]

The address is an identifier for the component that holds the choice